[DRUP-592] Adds a team prophecy trait.
